Add files via upload
These design files replace the functions of the 82284 and 82288 and provide shadow copy RAM. There are further enhancements to support 20MHz operation of the system. To use shadow RAM, this requires a section of machine code instructions to be added into an empty area inside your system ROM. I will document and include these additions.